Cast hours into integer

+++ templatetags/header.py
@@ -10,7 +10,7 @@ register = template.Library()
 def header():
     all_shows = Show.objects.filter(archived=False).order_by("category__name", "short")
 
-    hours_left = ceil((datetime(2015, 6, 23, 23, 59, 59) - datetime.now()).total_seconds() / 60 / 60)
+    hours_left = int(ceil((datetime(2015, 6, 23, 23, 59, 59) - datetime.now()).total_seconds() / 60 / 60))
 
     return {
         'all_shows': all_shows,
